requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
torch>=2.1.0
torch-geometric>=2.4.0
pydantic>=2.5.0
typer>=0.9.0
rich>=13.7.0
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
python-dotenv>=1.0.0
numpy>=1.24.0